+++ nodata/__init__.py
@@ -1,1 +1,1 @@
-from nodata.scripts import blob
+
