+++ tests/test_blob_cli.py
